Add files via upload
I have made available an Excel file specifically designed for data analysis purposes. It encompasses extensive data sets that can be explored and examined in-depth. In addition, to assist in the process of memorization, I have included succinct yet impactful sentences alongside the data. These sentences serve as mnemonic aids, allowing for easier retention and recall of the information.